pyproject.toml
[project]
name = "nonebot-plugin-gemini2"
version = "0.1.0"
description = "gemini聊天，支持文本，图片生成"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.10.0",
    "httpx>=0.28.1",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-alconna>=0.57.2",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-htmlrender>=0.6.3",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot-plugin-uninfo>=0.7.2",
    "nonebot-plugin-waiter>=0.8.1",
    "nonebot2[fastapi,httpx,websockets]>=2.4.2",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]

plugins = [
    "nonebot_plugin_alconna",
    "nonebot_plugin_uninfo",
    "nonebot_plugin_waiter",
    "nonebot_plugin_localstore",
    "nonebot_plugin_htmlrender",
    "nonebot_plugin_apscheduler",
]

plugin_dirs = ["."]

[dependency-groups]
dev = ["ruff>=0.11.5"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
ignore = ["E402"]
